requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
python-dotenv==1.0.1